requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
atlastk==0.13.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.6.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
